requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
envs==1.4
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.4.1
pydantic_core==2.10.1
pytailwindcss==0.2.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
